[Fix] Character file directory, annotation, data type
