--- src/main/java/donggrami/earth1round/src/character/CharacterRepository.java
@@ -1,9 +1,0 @@
-package donggrami.earth1round.src.character;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CharacterRepository extends JpaRepository<Character, Long> {
-    Character findByCharacterId(String name);
-}
